[PR #10672/35c979b backport][3.11] Fix headers being mutated if passed to web.Response as a CIMultiDict (#10673)
If `CIMultiDict` is passed in we need to make a copy to avoid mutating
it. In some cases we used to copy these twice which was fixed in #10045
but for this case that was the only copy being made and the source of
this regression.

fixes #10670

(cherry picked from commit 35c979be79dcc75ecbf7270ccc9fde264e5e2948)