Added a digest authentication helper (2/2)
This completes the implementation of the digest helper, implementing for
additional hashing algorithms sha256 and sha512, as well as auth-int
qop. It implements tests and documents the features.

The DigestAuth feature was shimmed into the existing API so it could be
used in place of BasicAuth in the auth field of ClientSession and
ClientRequest.