Support using system llhttp library (#10759)
Add a new `USE_SYSTEM_DEPS` environment variable that can be used
to enable building against the system install of llhttp library rather
than the vendored one.

This is implemented using the `pkgconfig` package to query llhttp
install via pkg-config.  The package is only used when `USE_SYSTEM_DEPS`
is enabled, however due to limitations of `pyproject.toml` it is listed
unconditionally as a build dependency.  That said, flake8 claims that
the dependency is not listed and I don't really know how to solve that:

```
setup.py:36:5: I900 'pkgconfig' not listed as a requirement
```

I have also updated `aiohttp/_cparser.pxd` to reference the `llhttp.h`
header via its name rather than full path.  I have poisoned the system
header to confirm that the correct header is used both when building
against the system library and the vendored library.