Reflect llhttp license in core packaging metadata
aiohttp vendors llhttp in its source distributions, and it also bundles it as a part for platform-specific wheels. Previously, this was not exposed in the core packaging metadata.

With this patch, now it is. The change includes both the license file from the vendored project and adds it to the SPDX expression following PEP 639.

This is configured through the `setup.cfg` config for the `setuptools` build backend.

PR #11226

Co-Authored-By: Karolina Surma <ksurma@redhat.com>
Co-Authored-By: 🇺🇦 Sviatoslav Sydorenko <webknjaz@redhat.com>